# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface AuthUser {
  id: number;
  username: string;
  fullName: string;
  role: string;
  branchId?: number | null;
  branchName?: string | null;
}

export interface User {
  id: number;
  username: string;
  fullName: string;
  role: string;
  branchId?: number | null;
  active: boolean;
  createdAt: string;
}

export interface CreateUserBody {
  username: string;
  password: string;
  fullName: string;
  role: string;
  branchId?: number | null;
}

export interface UpdateUserBody {
  fullName?: string;
  role?: string;
  branchId?: number | null;
  active?: boolean;
  password?: string;
}

export interface Branch {
  id: number;
  name: string;
  address?: string | null;
  phone?: string | null;
}

export interface CreateBranchBody {
  name: string;
  address?: string | null;
  phone?: string | null;
}

export interface Category {
  id: number;
  name: string;
}

export interface CreateCategoryBody {
  name: string;
}

export interface Supplier {
  id: number;
  name: string;
  contactName?: string | null;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
}

export interface CreateSupplierBody {
  name: string;
  contactName?: string | null;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
}

export interface Product {
  id: number;
  name: string;
  barcode?: string | null;
  sku?: string | null;
  categoryId?: number | null;
  categoryName?: string | null;
  supplierId?: number | null;
  supplierName?: string | null;
  branchId?: number | null;
  unit?: string | null;
  costPrice: number;
  sellPrice: number;
  stockQty: number;
  reorderLevel: number;
  expiryDate?: string | null;
  batchNumber?: string | null;
  description?: string | null;
}

export interface CreateProductBody {
  name: string;
  barcode?: string | null;
  sku?: string | null;
  categoryId?: number | null;
  supplierId?: number | null;
  branchId?: number | null;
  unit?: string | null;
  costPrice: number;
  sellPrice: number;
  stockQty: number;
  reorderLevel: number;
  expiryDate?: string | null;
  batchNumber?: string | null;
  description?: string | null;
}

export interface UpdateProductBody {
  name?: string;
  barcode?: string | null;
  sku?: string | null;
  categoryId?: number | null;
  supplierId?: number | null;
  branchId?: number | null;
  unit?: string | null;
  costPrice?: number;
  sellPrice?: number;
  stockQty?: number;
  reorderLevel?: number;
  expiryDate?: string | null;
  batchNumber?: string | null;
  description?: string | null;
}

export interface StockMovement {
  id: number;
  productId: number;
  productName: string;
  type: string;
  quantity: number;
  note?: string | null;
  userId?: number | null;
  userName?: string | null;
  createdAt: string;
}

export interface CreateStockMovementBody {
  productId: number;
  type: string;
  quantity: number;
  note?: string | null;
}

export interface SaleItem {
  id: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  costPrice: number;
  lineTotal: number;
}

export interface Sale {
  id: number;
  receiptNumber: string;
  subtotal: number;
  discount: number;
  tax: number;
  total: number;
  profit: number;
  paymentMethod: string;
  amountPaid: number;
  change: number;
  cashierId?: number | null;
  cashierName?: string | null;
  branchId?: number | null;
  branchName?: string | null;
  customerName?: string | null;
  createdAt: string;
  items: SaleItem[];
}

export interface CreateSaleItemBody {
  productId: number;
  quantity: number;
  unitPrice: number;
}

export interface CreateSaleBody {
  items: CreateSaleItemBody[];
  discount: number;
  tax: number;
  paymentMethod: string;
  amountPaid: number;
  customerName?: string | null;
}

export interface DashboardSummary {
  salesToday: number;
  salesWeek: number;
  salesMonth: number;
  profitToday: number;
  profitMonth: number;
  transactionsToday: number;
  productCount: number;
  lowStockCount: number;
  expiringSoonCount: number;
  totalStockValue: number;
}

export interface SalesTrendPoint {
  date: string;
  sales: number;
  profit: number;
}

export interface TopProduct {
  productId: number;
  productName: string;
  quantitySold: number;
  revenue: number;
}

export type ListProductsParams = {
  search?: string;
  barcode?: string;
  branchId?: number;
};

export type ListExpiringProductsParams = {
  days?: number;
};

export type ListStockMovementsParams = {
  productId?: number;
  limit?: number;
};

export type ListSalesParams = {
  limit?: number;
  from?: string;
  to?: string;
};

export type GetSalesTrendParams = {
  days?: number;
};

export type GetTopProductsParams = {
  limit?: number;
};