# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pharmacy Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
...
  /branches:
    get:
      operationId: listBranches
      tags: [branches]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBranchBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/{id}:
    patch:
      operationId: updateBranch
      tags: [branches]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBranchBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    delete:
      operationId: deleteCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: barcode
          in: query
          required: false
          schema:
            type: string
        - name: branchId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /products/expiring:
    get:
      operationId: listExpiringProducts
      tags: [products]
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
  /dashboard/sales-trend:
    get:
      operationId: getSalesTrend
      tags: [dashboard]
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesTrendPoint"
  /dashboard/top-products:
    get:
      operationId: getTopProducts
      tags: [dashboard]
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
        branchId:
          type: integer
          nullable: true
        branchName:
          type: string
          nullable: true
      required: [id, username, fullName, role]
...
    Branch:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
      required: [id, name]
    CreateBranchBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
      required: [name]
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required: [id, name]
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
      required: [name]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        barcode:
          type: string
          nullable: true
        sku:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        supplierId:
          type: integer
          nullable: true
        supplierName:
          type: string
          nullable: true
        branchId:
          type: integer
          nullable: true
        unit:
          type: string
          nullable: true
        costPrice:
          type: number
        sellPrice:
          type: number
        stockQty:
          type: integer
        reorderLevel:
          type: integer
        expiryDate:
          type: string
          nullable: true
        batchNumber:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
      required: [id, name, costPrice, sellPrice, stockQty, reorderLevel]
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        barcode:
          type: string
          nullable: true
        sku:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        supplierId:
          type: integer
          nullable: true
        branchId:
          type: integer
          nullable: true
        unit:
          type: string
          nullable: true
        costPrice:
          type: number
        sellPrice:
          type: number
        stockQty:
          type: integer
        reorderLevel:
          type: integer
        expiryDate:
          type: string
          nullable: true
        batchNumber:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
      required: [name, costPrice, sellPrice, stockQty, reorderLevel]
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        barcode:
          type: string
          nullable: true
        sku:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        supplierId:
          type: integer
          nullable: true
        branchId:
          type: integer
          nullable: true
        unit:
          type: string
          nullable: true
        costPrice:
          type: number
        sellPrice:
          type: number
        stockQty:
          type: integer
        reorderLevel:
          type: integer
        expiryDate:
          type: string
          nullable: true
        batchNumber:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
...
    SalesTrendPoint:
      type: object
      properties:
        date:
          type: string
        sales:
          type: number
        profit:
          type: number
      required: [date, sales, profit]
    TopProduct:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantitySold:
          type: integer
        revenue:
          type: number
      required: [productId, productName, quantitySold, revenue]