# lib/api-zod/src/generated/types/branch.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Branch {
  id: number;
  name: string;
  address?: string | null;
  phone?: string | null;
}