# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  name: string;
  barcode?: string | null;
  sku?: string | null;
  categoryId?: number | null;
  supplierId?: number | null;
  branchId?: number | null;
  unit?: string | null;
  costPrice: number;
  sellPrice: number;
  stockQty: number;
  reorderLevel: number;
  expiryDate?: string | null;
  batchNumber?: string | null;
  description?: string | null;
}