# lib/api-zod/src/generated/types/salesTrendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface SalesTrendPoint {
  date: string;
  sales: number;
  profit: number;
}